chore(parser): move discriminator file to v3 parser
