--- src/utils/isEqual.ts
@@ -1,40 +1,0 @@
-export const isEqual = <A, B>(a: A, b: B): boolean => {
-    // @ts-ignore
-    if (a === b) {
-        return true;
-    }
-
-    if (a && b && typeof a === 'object' && typeof b === 'object') {
-        if (Array.isArray(a) && Array.isArray(b)) {
-            if (a.length !== b.length) {
-                return false;
-            }
-            for (let i = 0; i < a.length; i++) {
-                if (!isEqual(a[i], b[i])) {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        const keysA = Object.keys(a);
-        const keysB = Object.keys(b);
-        if (keysA.length !== keysB.length) {
-            return false;
-        }
-
-        for (let i = 0; i < keysA.length; i++) {
-            const key = keysA[i];
-            if (!Object.prototype.hasOwnProperty.call(b, key)) {
-                return false;
-            }
-            // @ts-ignore
-            if (!isEqual(a[key], b[key])) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    return a !== a && b !== b;
-};
